Move new workspace creation to TransitionPage
