+++ src/libs/Navigation/AppNavigator/SharedScreens.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
-import TransitionPage from '../../../pages/Transition/TransitionPage';
-import TransitionNewWorkspace from '../../../pages/Transition/TransitionNewWorkspace';
+import TransitionPage from '../../../pages/TransitionPage';
 import SCREENS from '../../../SCREENS';
 import defaultScreenOptions from './defaultScreenOptions';
 
@@ -13,11 +12,6 @@ const SharedScreens = () => (
             name={SCREENS.TRANSITION}
             options={defaultScreenOptions}
             component={TransitionPage}
-        />
-        <RootStack.Screen
-            name="NewWorkspace"
-            options={defaultScreenOptions}
-            component={TransitionNewWorkspace}
         />
     </RootStack.Navigator>
 );
